refactor(gui): extract paths.rs inline seed_extract_tests into paths/tests.rs
paths.rs was 551 lines; its second inline `#[cfg(test)] mod
seed_extract_tests` block (the two write_seed_files corpus-fingerprint
tests) is moved verbatim into the already-extracted paths/tests.rs,
dropping paths.rs to 492 lines (under the 500-line guideline). The tests
now run as paths::tests::write_seed_files_* instead of
paths::seed_extract_tests::*; both still use `use super::*` so no import
change was needed (they use fully-qualified std::fs::). No behaviour
change; 207/207 primer-gui tests still pass.
